Safari-609.1.11.1/LayoutTests/fast/scrolling/ios/programmatic-scroll-via-scrollIntoView-should-clamp-top-expected.txt

This tests programming scrolling via element.scrollIntoView to an element position well above the viewport top
WebKit should clamp the scroll top to 0px.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS document.documentElement.scrollTop is 0
PASS visualViewport.pageTop is 0
PASS successfullyParsed is true

TEST COMPLETE
some content
